need newest changes to implement api calls

+++ backend/SwaggerRestApi/SwaggerRestApi/Models/Shelf.cs
@@ -6,8 +6,7 @@
 
         public int ShelfNo { get; set; }
 
-        public int Barcode { get; set; }
-
+        public Int64 Barcode { get; set; }
 
         public List<BaseItem> BaseItems { get; set; }
